# XDB-220 Enchanced the commit batching algorithm